fix: исправлены ошибки

+++ src/main/java/ru/yandex/practicum/filmorate/type/ReviewIdType.java
@@ -2,12 +2,12 @@ package ru.yandex.practicum.filmorate.type;
 
 import java.io.Serializable;
 
-public final class ReviewIdType extends ValueType<Integer> implements Serializable {
-    public ReviewIdType(Integer value) {
+public final class ReviewIdType extends ValueType<Long> implements Serializable {
+    public ReviewIdType(Long value) {
         super(value);
     }
 
-    public static ReviewIdType of(Integer value) {
+    public static ReviewIdType of(Long value) {
         return new ReviewIdType(value);
     }
 }
